updated and type casted the for loop to have a int cast

+++ Javascript/prime.py
@@ -4,7 +4,7 @@ def is_prime(n):
     """This function check if a number is prime or not"""
     if n < 2:
         return False
-    for i in range(2, math.sqrt(n)):
+    for i in range(2, int(math.sqrt(n))):
         if n % i == 0:
             return False
     return True
